# reduce memory use in HilbertCurve.toIndex 